chore: move expression back to column module, remove alias mixin
